Update the file structure
- Add Game class in lib/game.rb
- Move methods from main.rb to class Game
